# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Pagent API
  version: 0.1.0
  description: |
    Hosted UI rendering for terminal-bound AI agents. The agent emits an
    A2UI surface to this service, prints a short URL, and reads the user's
    interactions back via this API.

    Spec contents are opaque to the service. The `format` field on `POST /new`
    selects how the spec is interpreted:
    - `a2ui` (default) — `spec` is an A2UI v0.9 message array. Pages take user
      input and produce a result.
    - `html` — `spec` is a UTF-8 HTML string (up to 1,000,000 bytes).
      Pages are view-only: they never transition past `open`, and
      `POST /{id}/result` returns 400 `invalid_for_format`.

    See https://github.com/blockful/pagent for the source.

    ## Page lifecycle

    A page walks a 3-state machine: `open → submitted → received`.

    - **open** — created, waiting for a browser submission.
    - **submitted** — the browser has called `POST /{id}/result`. The agent
      has not yet read the result.
    - **received** — the agent called `GET /{id}/result` at least once after
      submission. The state flip is idempotent; subsequent reads keep returning
      `received`.

    ## Authentication modes

    When `REQUIRE_AUTH=false` (rollout grace mode), anonymous requests to
    `POST /new` and `GET /{id}/result` are allowed. OAuth scope checks still
    apply whenever a valid Bearer token is presented: an under-scoped token
...
  contact:
    name: Pagent maintainers
...
  license:
...
servers:
...
paths:
...
  /v1/share:
    get:
      operationId: getShareMetadata
      tags: [Viewer]
      summary: Resolve non-sensitive metadata and availability for a share token
      parameters: [{ $ref: '#/components/parameters/ShareToken' }]
      responses:
        {
          '200': { description: Deck title, sender, access mode, and state; never deck content. },
          '404': { description: Generic unavailable response. },
        }
  /v1/share/access:
    post:
      operationId: requestViewerSession
      tags: [Viewer]
      summary: Open a viewer session under the link access mode
      parameters: [{ $ref: '#/components/parameters/ShareToken' }]
      requestBody:
        content:
          {
            application/json:
              { schema: { type: object, properties: { email: { type: string, format: email } } } },
          }
      responses:
        {
          '200': { description: Granted session or generic access gate state. },
          '404': { description: Generic unavailable response. },
          '429': { description: Per-client-IP viewer write limit exceeded. },
        }
  /v1/share/request:
    post:
      operationId: createAccessRequest
      tags: [Viewer]
      summary: Ask the sender for link-specific access
      parameters: [{ $ref: '#/components/parameters/ShareToken' }]
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [email],
                    properties: { email: { type: string, format: email } },
                  },
              },
          }
      responses:
        {
          '202': { description: Generic request accepted response. },
          '404': { description: Generic unavailable response. },
          '429': { description: Per-client-IP viewer write limit exceeded. },
        }
  /v1/share/request/{requestId}:
    parameters:
      - { $ref: '#/components/parameters/ShareToken' }
      - { name: requestId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      operationId: getAccessRequestStatus
      tags: [Viewer]
      summary: Check a link-specific request state
      responses:
        {
          '200': { description: Pending, approved, or denied. },
          '404': { description: Generic unavailable response. },
        }
  /v1/viewer/deck:
    get:
      operationId: getViewerDeck
      tags: [Viewer]
      summary: Fetch sanitized deck content for a live viewer session
      parameters: [{ $ref: '#/components/parameters/ViewerSession' }]
      responses:
        {
          '200': { description: Ordered slides and identity confidence. },
          '410': { description: Session expired, revoked, or deck deleted. },
        }
  /v1/viewer/visits:
    post:
      operationId: startViewerVisit
      tags: [Engagement]
      summary: Start a human visit only after visibility and interaction qualification
      parameters: [{ $ref: '#/components/parameters/ViewerSession' }]
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/StartVisitRequest' } } }
      responses:
        {
          '200':
            {
              description: Visit started,
              excluded as preview or automation,
              or disabled by consent.,
            },
          '429': { description: Per-client-IP viewer write limit exceeded. },
        }
  /v1/viewer/visits/{visitId}/events:
    parameters:
      - { $ref: '#/components/parameters/ViewerSession' }
      - { name: visitId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      operationId: ingestEngagementEvents
      tags: [Engagement]
      summary: Idempotently ingest ordered or out-of-order engagement events
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/EngagementBatch' } } }
      responses:
        {
          '202': { description: Events accepted idempotently. },
          '410': { description: Session is unavailable or does not own the visit. },
          '429': { description: Per-client-IP engagement delivery limit exceeded. },
        }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OAuth access token. Bearer-authenticated calls to `POST /new` require
        `page:create`, and Bearer-authenticated calls to `GET /{id}/result`
        require `page:read`. These scope checks also apply when
        `REQUIRE_AUTH=false`; that mode permits anonymous calls, not
        under-scoped Bearer calls.
    CookieAuth:
      type: apiKey
      in: cookie
      name: pagent_session
      description: HttpOnly browser session cookie issued after interactive login.
  parameters:
...
    ShareToken:
      name: X-Share-Token
      in: header
      required: true
      description: Opaque token returned only when the share link is created.
      schema: { type: string, minLength: 20 }
    ViewerSession:
      name: X-Viewer-Session
      in: header
      required: true
      description: Opaque viewer or owner-preview session token with a maximum seven-day lifetime.
      schema: { type: string, minLength: 20 }
...
  schemas:
...
    StartVisitRequest:
      type: object
      additionalProperties: false
      required: [visible, interacted, analyticsConsent, deviceClass, browserFamily, countryCode]
      properties:
        visible:
          { type: boolean, description: True only when the viewer is at least 50 percent visible. }
        interacted: { type: boolean }
        analyticsConsent: { type: boolean }
        deviceClass: { type: string, enum: [mobile, tablet, desktop, unknown] }
        browserFamily: { type: string, maxLength: 80 }
        countryCode: { type: [string, 'null'], minLength: 2, maxLength: 2 }
    EngagementBatch:
      type: object
      additionalProperties: false
      required: [events]
      properties:
        events:
          type: array
          minItems: 1
          maxItems: 10
          items:
            type: object
            additionalProperties: false
            required:
              [
                id,
                eventType,
                eventAt,
                sequence,
                visibleRatio,
                visibleDurationMs,
                tabVisible,
                recentlyActive,
              ]
            properties:
              id: { type: string, format: uuid, description: Idempotency key. }
              eventType: { type: string, enum: [start, heartbeat, slide_view, close] }
              slideId: { type: string, format: uuid }
              eventAt: { type: string, format: date-time }
              sequence: { type: integer, minimum: 0 }
              visibleRatio: { type: number, minimum: 0, maximum: 1 }
              visibleDurationMs: { type: integer, minimum: 0, maximum: 60000 }
              tabVisible: { type: boolean }
              recentlyActive:
                {
                  type: boolean,
                  description: False after 60 seconds without mouse,
                  touch,
                  or keyboard activity.,
                }
...